+++ KaguyaProjectV2/KaguyaBot/Core/Services/GameRotationService.cs
@@ -31,7 +31,8 @@ namespace KaguyaProjectV2.KaguyaBot.Core.Services
                     new Tuple<string, ActivityType>($"{client.TotalUsers():N0} users", ActivityType.Watching),
                     new Tuple<string, ActivityType>($"{client.Guilds.Count:N0} servers", ActivityType.Watching),
                     new Tuple<string, ActivityType>($"$help | @Kaguya help", ActivityType.Listening),
-                    new Tuple<string, ActivityType>("$vote for bonuses", ActivityType.Watching)
+                    new Tuple<string, ActivityType>("$vote for bonuses", ActivityType.Watching),
+                    new Tuple<string, ActivityType>("$premium", ActivityType.Watching)
                 };
 
                 if (index >= games.Count)